requirements.txt
numpy==2.1.2
Pillow==11.0.0
datasets>=2.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
transformers>=4.37.0
accelerate>=0.20.0
peft>=0.8.0
sentencepiece>=0.1.99
protobuf>=3.20.0
opencv-python>=4.7.0
fire>=0.5.0
wandb>=0.15.0
bitsandbytes>=0.41.0